strategy

+++ src/main/java/com/example/studentmanagement/model/ViolationType.java
@@ -18,5 +18,5 @@ public class ViolationType {
     private String violationName;
 
     @Column(name = "deducted_points")
-    private Float deductedPoints;
+    private Double deductedPoints;
 }
